app: the OCR features has been removed from the app to make things simple
