+++ requirements.txt
@@ -71,9 +71,6 @@ mypy_extensions==1.1.0
 networkx==3.5
 numpy==2.3.0
 oauthlib==3.2.2
-openai==1.91.0
-opencv-contrib-python==4.10.0.84
-opencv-python==4.11.0.86
 openpyxl==3.1.5
 opt-einsum==3.3.0
 orjson==3.10.18
@@ -82,8 +79,6 @@ paddleocr==3.0.2
 paddlepaddle==3.0.0
 paddlex==3.0.2
 pandas==2.3.0
-pdf2image==1.17.0
-pillow==11.2.1
 premailer==3.10.0
 preshed==3.0.10
 prettytable==3.16.0
@@ -100,7 +95,6 @@ pydantic_core==2.33.2
 Pygments==2.19.1
 pyparsing==3.2.3
 pypdfium2==4.30.1
-pytesseract==0.3.13
 python-dateutil==2.9.0.post0
 python-dotenv==1.1.1
 pytz==2025.2
@@ -136,6 +130,7 @@ typing_extensions==4.14.0
 tzdata==2025.2
 ujson==5.10.0
 uritemplate==4.2.0
+urllib3==2.4.0
 wasabi==1.1.3
 wcwidth==0.2.13
 weasel==0.4.1
